pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "complypoint-phase1"
version = "0.1.0"
description = "Phase 1: Metadata Discovery & Readiness Engine"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1",
    "openpyxl>=3.1",
    "python-calamine>=0.2",
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
